# src/server-config.ts
import cors from 'cors';
import express, { json } from 'express';
import { moviesRoute } from './routes';

const server = express();
server.use(
  cors({
    origin: 'http://localhost:3000'
  })
);
server.use(json());

server.get('/', (_req, res) => {
  res.status(200).json({ message: 'Server is running!' });
});

server.use('/movies', moviesRoute);

server.get(
  '/auth/fake-token',
  (
    _req: express.Request,
    res: express.Response,
    next: express.NextFunction
  ) => {
    try {
      const token = `Bearer ${new Date().toISOString()}`;
      res.status(200).json({ token, status: 200 });
      return;
    } catch (error) {
      next(error);
    }
  }
);

export { server };